requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
python-dotenv==1.0.0
lxml==4.9.3
